bot/requirements.txt
python-telegram-bot==20.8
aiohttp==3.9.3
python-dotenv==1.0.1
